==> notebooks_data_prep/08_follow_up.ipynb <==
%load_ext autoreload
%autoreload 2

from collections import defaultdict

import dvu
import matplotlib.pyplot as plt
import pandas as pd
from os.path import join
import numpy as np
import seaborn as sns
import sys
sys.path.append('../notebooks_llm/')
plt.style.use('default')
dvu.set_style()
UNKS = {"Unk", "-", -1, "-1"}
df = pd.read_csv('../data/follow_up.csv')

df['Follow-up? (Y/N)'].value_counts()

df['Direct or Indirect?'].value_counts()
